# openapi/openapi.yml
openapi: 3.1.0
info:
  title: MLB Forecast API
  description: "\n    # MLB Predictions Backend API\n\n    Enterprise-grade MLB game\
    \ prediction platform delivering real-time analytics and machine learning insights.\n\
    \n    ## \U0001F680 Core Features\n\n    ### \u26BE **Data Ingestion Pipeline**\n\
    \    - Real-time game results from official MLB APIs\n    - Historical statistics\
    \ and performance metrics\n    - Live game schedules and league standings\n  \
    \  - Automated data validation and cleansing\n\n    ### \U0001F916 **Advanced\
    \ Machine Learning**\n    - Probabilistic game outcome predictions\n    - Total\
    \ runs and score forecasting\n    - Player performance analytics\n    - Continuous\
    \ model training and optimization\n\n    ### \u26A1 **High-Performance Architecture**\n\
    \    - Multi-layer Redis caching system\n    - Optimized database queries with\
    \ indexing\n    - Asynchronous processing for scalability\n    - Sub-second API\
    \ response times\n\n    ### \U0001F4CA **Analytics & Insights**\n    - Team performance\
    \ trending\n    - Head-to-head historical analysis\n    - Weather impact on game\
    \ outcomes\n    - Injury reports integration\n\n    ## \U0001F527 Technical Stack\n\
    \n    - **Framework**: FastAPI with async/await patterns\n    - **Database**:\
    \ PostgreSQL with optimized schemas\n    - **Cache**: Redis for high-speed data\
...
    \    ```\n    "
  version: 1.0.0
paths:
  /api/v1/teams:
    get:
      tags:
        - Teams
      summary: List Teams
      description: "Get a list of teams with optional filtering by league or division.\n\
        \nThis endpoint retrieves a list of teams. It supports filtering the teams\
        \ by\nleague or division. Filters are optional, and if not provided, all teams\
        \ will\nbe returned.\n\nArgs:\n    league: Optional filter by league (e.g.,\
        \ American or National)\n    division: Optional filter by division (e.g.,\
        \ East, West, Central)\n    use_cases: Dependency injected dictionary containing\
        \ use cases\n\nReturns:\n    JSONResponse: Standardized response with list\
        \ of teams\n\nRaises:\n    DomainExceptions.InvalidDataError: If invalid filter\
        \ parameters are provided"
      operationId: list_teams_api_v1_teams_get
      parameters:
        - name: league
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Filter by league (e.g. American or National)
            title: League
          description: Filter by league (e.g. American or National)
        - name: division
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Filter by division (e.g. East, West, Central)
            title: Division
          description: Filter by division (e.g. East, West, Central)
      responses:
        '200':
          description: Response successfully to retrieve teams list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/teams/{team_id}:
    get:
      tags:
        - Teams
      summary: Get Team
      description: "Get team details by team ID.\n\nThis endpoint retrieves the details\
        \ of a specific team using its unique identifier.\n\nArgs:\n    team_id: The\
        \ ID of the team to retrieve\n    use_cases: Dependency injected dictionary\
        \ containing use cases\n\nReturns:\n    JSONResponse: Standardized response\
        \ with team details\n\nRaises:\n    DomainExceptions.TeamNotFoundError: If\
        \ the team is not found\n    DomainExceptions.InvalidDataError: If team_id\
        \ is invalid"
      operationId: get_team_api_v1_teams__team_id__get
      parameters:
        - name: team_id
          in: path
          required: true
          schema:
            type: integer
            description: The ID of the team to get
            title: Team Id
          description: The ID of the team to get
      responses:
        '200':
          description: Successfully retrieved team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetailResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/teams:
    post:
      tags:
        - Teams
        - Data Ingestion
      summary: Ingest Teams
      description: "Ingest team data from external MLB API.\n\nThis endpoint triggers\
        \ the ingestion of team data from the official MLB API\nand returns metadata\
        \ about the operation.\n\nArgs:\n    use_cases: Dependency injected dictionary\
        \ containing use cases\n\nReturns:\n    JSONResponse: Standardized response\
        \ with ingestion results\n\nRaises:\n    DomainExceptions.ExternalServiceError:\
        \ If MLB API is unavailable"
      operationId: ingest_teams_api_v1_data_ingest_teams_post
      responses:
        '201':
          description: Successfully ingested team data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataIngestionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/teams/{team_id}/stats/{season}:
    get:
      tags:
        - Teams
        - Stats
      summary: Get Team Stats
      description: >
        Returns season-level composite statistics for a team, grouped by
        hitting, pitching, fielding and catching. Use the optional `category`
        query parameter to limit the response to a single block while the
        remaining categories are returned as `null` for compatibility.
      operationId: get_team_season_composite_stats
      parameters:
        - name: team_id
          in: path
          required: true
          schema:
            type: integer
            title: Team Id
            description: The ID of the team to get stats for
          description: The ID of the team to get stats for
        - name: season
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum:
              - hitting
              - pitching
              - fielding
              - catching
              - all
            default: all
          description: "Optional stats category filter. Defaults to `all`."
      responses:
        '200':
          description: Successfully retrieved team composite statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamSeasonStatsDetailResponse'
              examples:
                all:
                  summary: All categories
                  value:
                    status: success
                    code: 200
                    data:
                      team_id: 62
                      season: 2025
                      hitting:
                        hits: 1450
                        games_played: 162
                      pitching:
                        wins: 90
                        games_played: 162
                      fielding:
                        total_chances: 5100
                        games_played: 162
                      catching:
                        caught_stealing: 45
                        games_played: 162
                      updated_at: '2025-07-25T17:58:28Z'
                    errors: []
                    message: Team statistics retrieved successfully
                hitting:
                  summary: Hitting category only
                  value:
                    status: success
                    code: 200
                    data:
                      team_id: 62
                      season: 2025
                      hitting:
                        hits: 1450
                        games_played: 162
                      pitching: null
                      fielding: null
                      catching: null
                      updated_at: '2025-07-25T17:58:28Z'
                    errors: []
                    message: Team hitting statistics retrieved successfully
                pitching:
                  summary: Pitching category only
                  value:
                    status: success
                    code: 200
                    data:
                      team_id: 62
                      season: 2025
                      hitting: null
                      pitching:
                        wins: 90
                        innings_pitched: 1450.1
                      fielding: null
                      catching: null
                      updated_at: '2025-07-25T17:58:28Z'
                    errors: []
                    message: Team pitching statistics retrieved successfully
                fielding:
                  summary: Fielding category only
                  value:
                    status: success
                    code: 200
                    data:
                      team_id: 62
                      season: 2025
                      hitting: null
                      pitching: null
                      fielding:
                        total_chances: 5100
                        fielding_percentage: 0.972
                      catching: null
                      updated_at: '2025-07-25T17:58:28Z'
                    errors: []
                    message: Team fielding statistics retrieved successfully
                catching:
                  summary: Catching category only
                  value:
                    status: success
                    code: 200
                    data:
                      team_id: 62
                      season: 2025
                      hitting: null
                      pitching: null
                      fielding: null
                      catching:
                        caught_stealing: 45
                        passed_balls: 5
                      updated_at: '2025-07-25T17:58:28Z'
                    errors: []
                    message: Team catching statistics retrieved successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
...
components:
  schemas:
...
    DataIngestionResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/IngestionSummaryData'
          description: Ingestion operation details
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: DataIngestionResponse
      description: Response model for data ingestion endpoints.
      content:
        application/json:
          example:
            status: success
            code: 201
            data:
              ingestion_summary:
                operation: team_ingestion
                records_processed: 30
                records_created: 30
                records_updated: 0
                errors: [ ]
                duration_seconds: 2.45
                timestamp: '2025-07-25T18:30:15.123456Z'
              sample_teams:
                - id: 1
                  mlb_id: 147
                  name: New York Yankees
                  abbreviation: NYY
                  city: New York
                  division: American League East
                  league: American League
                  venue_name: Yankee Stadium
                  created_at: '2025-07-25T17:58:28.647662Z'
            errors: [ ]
            message: Successfully ingested 30 teams
...
    BadRequest:
      description: Bad Request - Invalid input parameters
      title: Bad Request
      $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - Resource not found
      title: Not Found
      $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation Error - Request validation failed
      title: Validation Error
      $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      title: Internal Server Error
      $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service Unavailable - External service error
      title: Service Unavailable
      $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      properties:
        status:
          type: string
          title: Status
          description: Response status
          default: error
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          type: 'null'
          title: Data
          description: No data in error responses
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
          description: Error message
      type: object
      required:
        - code
        - errors
      title: ErrorResponse
      description: Error response model for API documentation.
      examples:
        - code: 400
          errors:
            - League must be 'American' or 'National'
          message: Invalid data provided
          status: error
        - code: 404
          errors:
            - Team with ID 999 not found
          message: Resource not found
          status: error
        - code: 500
          errors:
            - An unexpected error occurred
          message: Internal server error
          status: error
...
    TeamDTO:
      properties:
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
          description: Internal team ID
        mlb_id:
          type: integer
          title: Mlb Id
          description: Official MLB team ID
        name:
          type: string
          title: Name
          description: Team name
        abbreviation:
          type: string
          title: Abbreviation
          description: Team abbreviation (e.g., NYY)
        city:
          type: string
          title: City
          description: Team city
        division:
          type: string
          title: Division
          description: Division (East, Central, West)
        league:
          type: string
          title: League
          description: League (American, National)
        venue_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Venue Name
          description: Home venue name
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Created At
          description: Record creation timestamp
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
          description: Record last update timestamp
      type: object
      required:
        - mlb_id
        - name
        - abbreviation
        - city
        - division
        - league
      title: TeamDTO
      description: Data Transfer Object for Team entity.
    TeamDetailResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/TeamDTO'
          description: Team details
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: TeamDetailResponse
      description: Response model for single team endpoints.
      example:
        status: success
        code: 200
        data:
          id: 1
          mlb_id: 147
          name: New York Yankees
          abbreviation: NYY
          city: New York
          division: American League East
          league: American League
          venue_name: Yankee Stadium
          created_at: '2025-07-25T17:58:28.647662Z'
        errors: [ ]
        message: Team New York Yankees retrieved successfully
    TeamListResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          items:
            $ref: '#/components/schemas/TeamDTO'
          type: array
          title: Data
          description: List of teams
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: TeamListResponse
      description: Response model for team list endpoints.
      example:
        status: success
        code: 200
        data:
          - id: 1
            mlb_id: 133
            name: Athletics
            abbreviation: ATH
            city: Sacramento
            division: American League West
            league: American League
            venue_name: Sutter Health Park
            created_at: '2025-07-25T17:58:28.639243Z'
          - id: 2
            mlb_id: 136
            name: Seattle Mariners
            abbreviation: SEA
            city: Seattle
            division: American League West
            league: American League
            venue_name: T-Mobile Park
            created_at: '2025-07-25T17:58:28.656583Z'
        errors: [ ]
        message: Retrieved 2 team successfully
...
    TeamSeasonStatsDetailResponse:
      type: object
      description: Response model for composite team stats.
      properties:
        status: { type: string, default: success }
        code: { type: integer, description: HTTP status code }
        data:
          $ref: '#/components/schemas/TeamSeasonStatsDTO'
        errors:
          type: array
          items: { type: string }
        message: { type: string }
      required: [ code, data ]
      example:
        status: success
        code: 200
        data:
          team_id: 62
          season: 2025
          hitting:
            games_played: 162
            plate_appearances: 6200
            at_bats: 5600
            runs: 780
            hits: 1450
            singles: 950
            doubles: 310
            triples: 25
            home_runs: 165
            rbi: 740
            walks: 520
            strikeouts: 1350
            stolen_bases: 120
            caught_stealing: 35
            hit_by_pitch: 75
            sacrifice_hits: 20
            sacrifice_flies: 45
            ground_into_double_play: 105
            total_bases: 2300
            batting_average: 0.259
            on_base_percentage: 0.325
            slugging_percentage: 0.410
            ops: 0.735
            stolen_base_percentage: 0.774
          pitching:
            games_played: 162
            games_started: 162
            wins: 89
            losses: 73
            saves: 42
            blown_saves: 12
            holds: 68
            innings_pitched: 1448.1
            batters_faced: 6100
            hits_allowed: 1330
            runs_allowed: 680
            earned_runs: 620
            home_runs_allowed: 180
            walks_allowed: 510
            strikeouts: 1485
            hit_batters: 58
            wild_pitches: 45
            balks: 3
            ground_outs: 1650
            air_outs: 1500
            era: 3.86
            whip: 1.23
            strikeouts_per_9: 9.2
            walks_per_9: 3.2
            home_runs_per_9: 1.1
            strikeout_walk_ratio: 2.9
            opponent_batting_average: 0.242
          fielding:
            games_played: 162
            innings: 14250
            total_chances: 6150
            putouts: 4300
            assists: 1650
            errors: 200
            double_plays: 145
            triple_plays: 0
            fielding_percentage: 0.967
            range_factor_per_game: 5.9
          catching:
            games_caught: 162
            stolen_bases_allowed: 125
            caught_stealing: 45
            caught_stealing_percentage: 0.265
            passed_balls: 7
            wild_pitches: 42
            pickoffs: 6
            catcher_interference: 1
        errors: [ ]
        message: Team composite statistics retrieved successfully
...